Comment out most of the range feedback mechanism (going to rewrite). Continue refactoring the regression tests - look nice and clean now.
svn path=/trunk/KDE/kdelibs/kate/; revision=467779

+++ part/part/katelinerange.cpp
@@ -177,7 +177,7 @@ KateTextLayout KateLineLayout::viewLine( int viewLine ) const
 
 int KateLineLayout::width( ) const
 {
-  return layout()->boundingRect().width();
+  return (int)layout()->boundingRect().width();
 }
 
 bool KateLineLayout::isOutsideDocument( ) const
